\name{epiclomal.npm.cluster-package}
\alias{epiclomal.npm.cluster-package}
\alias{epiclomal.npm.cluster}
\docType{package}
\title{
  Non-probabilistic Methods for Epiclomal
}
\description{
  Non-probabilistic clustering methods to cluster methylation data. Used by the epiclomal workflow.
}
\details{
  Clustering methods include EuclideanClust, HammingClust (PBAL), PearsonClust, DensityCut
}
\author{
Camila P.E. de Souza, Mirela Andronescu

Maintainer: Patricia Ye <pye@bccrc.ca>
}
\references{
  de Souza CPE, Andronescu M, Masud T, Kabeer F, Biele J, Laks E, et al. Epiclomal: probabilistic clustering of sparse single-cell DNA methylation data. bioRxiv. 2018;414482.
}
% \keyword{ package }
% \seealso{
%   Optional links to other man pages
% }
% \examples{
%   \dontrun{
%      ## Optional simple examples of the most important functions
%      ## These can be in \dontrun{} and \donttest{} blocks.
%   }
% }